requirements.txt
python-dotenv
langchain
langchain_community
faiss-cpu
langchain-openai
langgraph
beautifulsoup4
requests
sentence-transformers
pymilvus
asyncio
unstructured
pdfminer.six==20250506
pi_heif
pdfminer.psexceptions
pdf2image
pytesseract
PyMuPDF
neo4j
leidenalg
igraph
